Rename connection interface and split the definition from adb
