+++ source/madbfs.cpp
@@ -1,5 +1,6 @@
 #include "madbfs/madbfs.hpp"
 #include "madbfs/args.hpp"
+#include "madbfs/data/adb_connection.hpp"
 #include "madbfs/data/ipc.hpp"
 #include "madbfs/log.hpp"
 #include "madbfs/util/overload.hpp"
@@ -93,7 +94,7 @@ namespace madbfs
 {
     Madbfs::Madbfs(usize page_size, usize max_pages)
         : m_work_guard{ m_async_ctx.get_executor() }
-        , m_connection{ std::make_unique<data::Connection>() }
+        , m_connection{ std::make_unique<data::AdbConnection>() }
         , m_cache{ page_size, max_pages }
         , m_tree{ *m_connection, m_cache }
     {
